requirements.txt
Django==1.6.5
MySQL-python==1.2.5
South==1.0
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
feedparser==5.1.3
gunicorn==19.0.0
parse-rest==0.2.20141004
psycopg2==2.5.3
pystache==0.5.4
python-dateutil==2.2
six==1.7.3
static==1.0.2
static3==0.5.1
virtualenv==1.11.6
wsgiref==0.1.2
